requirements.txt
Flask==2.3.3
Jinja2==3.1.6
MarkupSafe==3.0.3
Pillow==12.0.0
Werkzeug==2.3.7
blinker==1.9.0
click==8.3.0
fpdf==1.7.2
gunicorn==21.2.0
itsdangerous==2.2.0
packaging==25.0
pytesseract==0.3.10
pdfplumber==0.10.3
pdf2image==1.16.3